removed old report view and replaced with new report view.
- added documentation to explain report node features and usage.
-remove old report view.
