+++ ApsimNG/Views/ReportView.cs
@@ -7,22 +7,28 @@ namespace UserInterface.Views
 {
 
     /// <summary>
-    /// View for a report component.
+    /// View for a report component that includes new report variable and report frequency UI sections.
     /// </summary>
     public class ReportView : ViewBase, IReportView
     {
+
         private Notebook notebook1 = null;
-        private VBox vbox1 = null;
-        private VBox vbox2 = null;
+        private Paned reportVariablesBox = null;
+        private Paned reportFrequencyBox = null;
+        private Box variablesBox = null;
+        private Box commonVariablesBox = null;
+        private Box frequencyBox = null;
+        private Box commonFrequencyBox = null;
         private Alignment alignment1 = null;
 
         private IEditorView variableEditor;
         private IEditorView frequencyEditor;
-        private IListView commonReportVariablesList;
-        private IListView commonReportFrequencyVariablesList;
+        private IListView commonReportVariableList;
+        private IListView commonReportFrequencyVariableList;
         private ViewBase dataStoreView1;
         private VPaned panel;
         private EditView groupByEdit;
+        private Button submitButton;
 
         /// <summary>
         /// Invoked when the user moves the vertical splitter
@@ -38,11 +44,17 @@ namespace UserInterface.Views
         /// <summary>Constructor</summary>
         public ReportView(ViewBase owner) : base(owner)
         {
-            Builder builder = BuilderFromResource("ApsimNG.Resources.Glade.ReportView.glade");
+            Builder builder = BuilderFromResource("ApsimNG.Resources.Glade.NewReportView.glade");
             notebook1 = (Notebook)builder.GetObject("notebook1");
-            vbox1 = (VBox)builder.GetObject("vbox1");
-            vbox2 = (VBox)builder.GetObject("vbox2");
+            reportVariablesBox = (Paned)builder.GetObject("reportVariablesBox");
+            variablesBox = (Box)builder.GetObject("variablesBox");
+            commonVariablesBox = (Box)builder.GetObject("commonVariablesBox");
+            reportFrequencyBox = (Paned)builder.GetObject("reportFrequencyBox");
+            frequencyBox = (Box)builder.GetObject("frequencyBox");
+            commonFrequencyBox = (Box)builder.GetObject("commonFrequencyBox");
             alignment1 = (Alignment)builder.GetObject("alignment1");
+            submitButton = (Button)builder.GetObject("submitBtn");
+
 
             panel = (VPaned)builder.GetObject("vpaned1");
             panel.Events |= Gdk.EventMask.PropertyChangeMask;
@@ -54,18 +66,37 @@ namespace UserInterface.Views
             mainWidget = notebook1;
             notebook1.SwitchPage += OnSwitchPage;
 
+            reportVariablesBox.AddNotification(OnVariablesPanePropertyNotified);
+            reportFrequencyBox.AddNotification(OnFrequencyPanePropertyNotified);
+
             variableEditor = new EditorView(this);
             variableEditor.StyleChanged += OnStyleChanged;
-            vbox1.PackStart((variableEditor as ViewBase).MainWidget, true, true, 0);
+            variablesBox.PackStart((variableEditor as ViewBase).MainWidget, true, true, 0);
 
             frequencyEditor = new EditorView(this);
             frequencyEditor.StyleChanged += OnStyleChanged;
-            vbox2.PackStart((frequencyEditor as ViewBase).MainWidget, true, true, 0);
+            frequencyBox.PackStart((frequencyEditor as ViewBase).MainWidget, true, true, 0);
+
+            commonReportVariableList = new ListView(this, new Gtk.TreeView(), new Gtk.Menu(), (EditorView)variableEditor);
+            commonReportVariableList.DoubleClicked += OnCommonReportVariableListDoubleClicked;
+            commonVariablesBox.PackStart((commonReportVariableList as ViewBase).MainWidget, true, true, 0);
+
+            commonReportFrequencyVariableList = new ListView(this, new Gtk.TreeView(), new Gtk.Menu(), (EditorView)frequencyEditor, submitButton);
+            commonReportFrequencyVariableList.DoubleClicked += OnCommonReportFrequencyVariableListDoubleClicked;
+            commonFrequencyBox.PackStart((commonReportFrequencyVariableList as ViewBase).MainWidget, true, true, 0);
+
+            // Set the position of the divider to 80% of the width of the Paned gtk object.
+            reportVariablesBox.Position = (int)Math.Round(this.owner.MainWidget.AllocatedWidth * 0.7);
+            reportFrequencyBox.Position = (int)Math.Round(this.owner.MainWidget.AllocatedWidth * 0.7);
 
             dataStoreView1 = new ViewBase(this, "ApsimNG.Resources.Glade.DataStoreView.glade");
             alignment1.Add(dataStoreView1.MainWidget);
             mainWidget.Destroyed += _mainWidget_Destroyed;
+
         }
+
+
+
 
         /// <summary>
         /// Invoked when the selected tab is changed.
@@ -87,6 +118,22 @@ namespace UserInterface.Views
             {
                 ShowError(err);
             }
+        }
+
+        /// <summary> Updates The position of either common variable listView.</summary>
+        /// <param name="sender"></param>
+        /// <param name="args"></param>
+        private void OnFrequencyPanePropertyNotified(object sender, NotifyArgs args)
+        {
+            this.reportVariablesBox.Position = reportFrequencyBox.Position;
+        }
+
+        /// <summary> Updates The position of either common variable listView.</summary>
+        /// <param name="sender"></param>
+        /// <param name="args"></param>
+        private void OnVariablesPanePropertyNotified(object sender, NotifyArgs args)
+        {
+            this.reportFrequencyBox.Position = reportVariablesBox.Position;
         }
 
         /// <summary>
@@ -126,6 +173,7 @@ namespace UserInterface.Views
             {
                 variableEditor?.Refresh();
                 frequencyEditor?.Refresh();
+
             }
             catch (Exception err)
             {
@@ -157,6 +205,16 @@ namespace UserInterface.Views
             }
         }
 
+        private void OnCommonReportVariableListDoubleClicked(object sender, EventArgs e)
+        {
+
+        }
+
+        private void OnCommonReportFrequencyVariableListDoubleClicked(object sender, EventArgs e)
+        {
+
+        }
+
         /// <summary>Provides access to the variable list.</summary>
         public IEditorView VariableList { get { return variableEditor; } }
 
@@ -166,11 +224,12 @@ namespace UserInterface.Views
         /// <summary>Provides access to the group by edit.</summary>
         public IEditView GroupByEdit { get { return groupByEdit; } }
 
+        public IListView CommonReportVariablesList { get { return commonReportVariableList; } set { commonReportVariableList = value; } }
+
+        public IListView CommonReportFrequencyVariablesList { get { return commonReportFrequencyVariableList; } set { commonReportFrequencyVariableList = value; } }
+
         /// <summary>Provides access to the DataGrid.</summary>
         public ViewBase DataStoreView { get { return dataStoreView1; } }
-
-        public IListView CommonReportVariablesList { get { return commonReportVariablesList; } set { commonReportVariablesList = value; } }
-        public IListView CommonReportFrequencyVariablesList { get { return commonReportFrequencyVariablesList; } set { commonReportFrequencyVariablesList = value; } }
 
         /// <summary>
         /// Indicates the index of the currently active tab
@@ -201,3 +260,4 @@ namespace UserInterface.Views
     }
 
 }
+
